device/allocator, removing the annoying windows hook, pattern matching fixes and more
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=1c023e44e4f1932f1cb7c6978c0ab3e737a6dac8

+++ components/profiles/src/IdentityProviderSteam.cpp
@@ -97,7 +97,7 @@ concurrency::task<ProfileIdentityResult> SteamIdentityProvider::ProcessIdentity(
 	*(uint16_t*)&ticketBuffer[8] = ticketLength;
 
 	// encode the ticket buffer
-	uint32_t outLength;
+	size_t outLength;
 	char* str = base64_encode(ticketBuffer, ticketLength + 10, &outLength);
 
 	// and return a result
